backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-huggingface
chromadb
sentence-transformers
groq
pypdf
python-dotenv
pydantic
python-multipart
